Make ProfileEvents.h less hot
Before: 2700TUs on ProfileEvents.h changes
After (as of now): 1234TUs
