+++ ci/jobs/scripts/check_style/check_cpp.sh
@@ -133,9 +133,12 @@ EXTERN_TYPES_EXCLUDES=(
 # and this matches with zkutil::CreateMode
 grep -v -e 'src/Common/ZooKeeper/Types.h' -e 'src/Coordination/KeeperConstants.cpp' "$STYLE_TMPDIR/all_excluded" > "$STYLE_TMPDIR/extern_files"
 
-# Extract declarations: "filepath:D TYPE NAME"
+# Extract declarations: "filepath:D TYPE NAME [NOLINT]"
+# A trailing NOLINT comment marks the declaration as used elsewhere: it still counts as a
+# declaration (so same-file usages and duplicates are checked), but is exempt from the
+# "defined but not used" check below.
 xargs < "$STYLE_TMPDIR/extern_files" rg -o --no-line-number \
-    'extern const (int|Event|Metric) ([_A-Za-z0-9]+);' -r 'D $1 $2' > "$STYLE_TMPDIR/extern_combined"
+    'extern const (int|Event|Metric) ([_A-Za-z0-9]+);(?:.*?(NOLINT))?' -r 'D $1 $2 $3' > "$STYLE_TMPDIR/extern_combined"
 
 # Extract usages (skipping comment lines): "filepath:U NS NAME"
 xargs < "$STYLE_TMPDIR/extern_files" rg --no-line-number \
@@ -172,6 +175,7 @@ BEGIN {
         ns = type_to_ns[p[2]]
         key = file SUBSEP ns SUBSEP p[3]
         decl[key]++
+        if (p[4] == "NOLINT") used[key] = 1
     } else {
         key = file SUBSEP p[2] SUBSEP p[3]
         used[key] = 1
