+++ src/Backups/BackupImpl.h
@@ -6,6 +6,7 @@
 #include <Backups/IBackupCoordination.h>
 #include <Backups/BackupInfo.h>
 #include <Common/Logger_fwd.h>
+#include <atomic>
 #include <map>
 #include <mutex>
