+++ src/Common/Arena.h
@@ -6,7 +6,6 @@
 #include <Core/Defines.h>
 #include <boost/noncopyable.hpp>
 #include <Common/Allocator.h>
-#include <Common/ProfileEvents.h>
 #include <Common/memcpySmall.h>
 #include <base/getPageSize.h>
 
@@ -14,12 +13,6 @@
 #   include <sanitizer/asan_interface.h>
 #endif
 
-
-namespace ProfileEvents
-{
-    extern const Event ArenaAllocChunks;
-    extern const Event ArenaAllocBytes;
-}
 
 namespace DB
 {
@@ -68,17 +61,7 @@ private:
             return *this;
         }
 
-        explicit MemoryChunk(size_t size_)
-        {
-            ProfileEvents::increment(ProfileEvents::ArenaAllocChunks);
-            ProfileEvents::increment(ProfileEvents::ArenaAllocBytes, size_);
-
-            begin = reinterpret_cast<char *>(Allocator<false>::alloc(size_));
-            pos = begin;
-            end = begin + size_ - pad_right;
-
-            ASAN_POISON_MEMORY_REGION(begin, size_);
-        }
+        explicit MemoryChunk(size_t size_);
 
         ~MemoryChunk()
         {
