+++ src/Common/PageCache.h
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <IO/BufferWithOwnMemory.h>
 #include <Common/CacheBase.h>
 #include <Common/CacheLine.h>
 #include <Common/HashTable/Hash.h>
