+++ src/Common/ProfileEvents.h
@@ -5,7 +5,6 @@
 #include <Interpreters/Context_fwd.h>
 #include <base/types.h>
 #include <base/strong_typedef.h>
-#include <Poco/Message.h>
 #include <atomic>
 #include <memory>
 #include <cstddef>
@@ -201,12 +200,6 @@ namespace ProfileEvents
     /// The same as above but ignores value of setting 'trace_profile_events'
     /// and never sends profile event to trace log.
     void incrementNoTrace(Event event, Count amount = 1);
-
-    /// Increment a counter for log messages.
-    void incrementForLogMessage(Poco::Message::Priority priority);
-
-    /// Increment time consumed by logging.
-    void incrementLoggerElapsedNanoseconds(UInt64 ns);
 
     /// Get name of event by identifier. Returns statically allocated string.
     const std::string_view & getName(Event event);
