+++ src/Common/Scheduler/CPUSlotsAllocation.h
@@ -12,6 +12,7 @@
 #include <atomic>
 #include <condition_variable>
 #include <mutex>
+#include <vector>
 
 namespace DB
 {
