+++ src/Common/ThreadPoolTaskTracker.cpp
@@ -1,7 +1,6 @@
-#include "config.h"
-
 #include <Common/MemoryTracker.h>
 #include <Common/ThreadPoolTaskTracker.h>
+#include <Common/ProfileEvents.h>
 
 namespace ProfileEvents
 {
