+++ src/Common/ThreadStatus.h
@@ -17,6 +17,7 @@
 #include <functional>
 #include <mutex>
 #include <unordered_set>
+#include <vector>
 
 
 template <class T>
