+++ src/IO/ReadBufferFromPocoSocket.h
@@ -4,6 +4,7 @@
 #include <IO/ReadBuffer.h>
 #include <Common/AsyncTaskExecutor.h>
 #include <Common/Stopwatch.h>
+#include <Common/ProfileEvents.h>
 #include <Poco/Net/Socket.h>
 
 namespace DB
