+++ src/IO/WriteBufferFromFileDescriptorDiscardOnFailure.cpp
@@ -1,4 +1,5 @@
 #include <IO/WriteBufferFromFileDescriptorDiscardOnFailure.h>
+#include <Common/ProfileEvents.h>
 
 namespace ProfileEvents
 {
